+++ .gitignore
@@ -19,3 +19,4 @@ carp/pytorch/data/utils/api_key.txt
 
 pytorch_model.bin
 config.json
+carp/data.zip
