+++ carp/pytorch/model/architectures/carp_coop.py
@@ -7,7 +7,7 @@ from carp.configs import ModelConfig
 from carp.pytorch.data.scarecrow_pipeline import ScarecrowTargetElement
 from carp.pytorch.model.architectures import *
 from carp.pytorch.model.encoders import BaseEncoder, BaseEncoderOutput
-from carp.pytorch.training import BaseTrainer, register_trainer
+from carp.pytorch.training.trainer import BaseTrainer, register_trainer
 from carp.util import generate_indices
 
 os.environ["TOKENIZERS_PARALLELISM"] = "true"
